moving telemetry patch/mock to cover full test suite
