+++ test/conftest.py
@@ -11,9 +11,12 @@ from unittest.mock import Mock
 
 import pytest
 
+import haystack.core.pipeline.pipeline as pipeline_module
+import haystack.telemetry._telemetry as telemetry_module
 from haystack import component, tracing
 from haystack.core.serialization import allow_deserialization_module
 from haystack.document_stores.in_memory import InMemoryDocumentStore
+from haystack.telemetry._telemetry import pipeline_running, send_telemetry
 from haystack.testing.test_utils import set_all_seeds
 from test.tracing.utils import EagerSpyingTracer, SpyingTracer
 
@@ -94,6 +97,43 @@ def request_blocker(request: pytest.FixtureRequest, monkeypatch):
         raise RuntimeError(f"The test was about to {method} {self.scheme}://{self.host}{url}")
 
     monkeypatch.setattr("urllib3.connectionpool.HTTPConnectionPool.urlopen", urlopen_mock)
+
+
+@pytest.fixture(autouse=True)
+def block_telemetry_network_calls(monkeypatch):
+    """
+    Force `telemetry` to be a real, truthy Telemetry instance (regardless of the ambient
+    HAYSTACK_TELEMETRY_ENABLED setting) so every Pipeline.run()/run_async() in the test suite
+    (directly, or indirectly via SuperComponent, PipelineTool, Agent, retrievers, etc.) still
+    exercises the real @send_telemetry / Telemetry.send_event() code path - catching any real
+    breakage there - but stub out `posthog.capture`, the only place that code path ever touches
+    the network, so no test ever actually sends data or connects to PostHog.
+    """
+    monkeypatch.setattr(telemetry_module, "telemetry", telemetry_module.Telemetry())
+    monkeypatch.setattr(telemetry_module.posthog, "capture", Mock())
+
+
+@pytest.fixture(autouse=True)
+def tag_pipeline_telemetry_as_test(monkeypatch):
+    """
+    Pipeline.run()/run_async() report a "Pipeline run (3.x)" event to Posthog on every call. Tag
+    events triggered by the test suite with a distinct name so they can be told apart from real
+    usage, without changing haystack.telemetry.pipeline_running itself.
+
+    `pipeline_running.__wrapped__` is the raw, undecorated function stashed there by
+    `functools.wraps` inside the `@send_telemetry` decorator - reusing it (instead of
+    reimplementing its logic here) means this stays in sync with the real function automatically.
+    """
+
+    @send_telemetry
+    def test_pipeline_running(pipeline):
+        result = pipeline_running.__wrapped__(pipeline)
+        if result is None:
+            return None
+        event_name, event_properties = result
+        return f"{event_name} (tests)", event_properties
+
+    monkeypatch.setattr(pipeline_module, "pipeline_running", test_pipeline_running)
 
 
 @pytest.fixture()
